+++ employee-service/src/main/java/com/djcodes/test/containers/crudservice/persistence/EmployeeRepository.java
@@ -2,7 +2,7 @@ package com.djcodes.test.containers.crudservice.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<EmployeeEntity, Long> {
+public interface EmployeeRepository extends JpaRepository<EmployeeEntity, Integer> {
 
     public EmployeeEntity findByFname(String fname);
